# swagger.json
{"openapi":"3.0.1","info":{"title":"API de HelpDesk","description":"API para gestionar tickets de soporte","contact":{"name":"Soporte Técnico","email":"soporte@empresa.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Ticket Controller","description":"Operaciones CRUD para tickets"}],"paths":{"/api/tickets/{id}":{"get":{"tags":["Ticket Controller"],"summary":"Obtener un ticket por ID","description":"Devuelve los detalles de un ticket específico","operationId":"getTicketById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Éxito, ticket encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}},"404":{"description":"Error, ticket no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}},"put":{"tags":["Ticket Controller"],"summary":"Editar un ticket por ID","description":"Edita un ticket especifico por la base de datos","operationId":"updateTicket","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}},"delete":{"tags":["Ticket Controller"],"summary":"Elimina un ticket por ID","description":"Elimina un ticket espefifco por la base de datos","operationId":"deleteTicket","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/tickets/":{"get":{"tags":["Ticket Controller"],"summary":"Busca todos los tickets","description":"devuelve todo los tickets guardados en la base de datos","operationId":"getAllTickets","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Ticket"}}}}}}},"post":{"tags":["Ticket Controller"],"summary":"Crear un nuevo ticket","description":"Guarda un nuevo ticket en la base de datos","operationId":"createTicket","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Ticket"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ticket"}}}}}}}},"components":{"schemas":{"Ticket":{"type":"object","properties":{"id":{"type":"string","description":"ID del ticket","example":"67a588fce93d0f71abcbe50f"},"title":{"type":"string","description":"Titulo del problema","example":"Problema con Correo"},"description":{"type":"string","description":"Descripción del problema","example":"No funciona el correo"},"status":{"type":"string","description":"Estado del problema","example":"Abierto"},"priority":{"type":"string","description":"Prioridad del problema","example":"Alta"},"assignedTo":{"type":"string"},"createdBy":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}},"description":"Entidad que representa un ticket de soporte"}}}}